+++ test/unit/strats.unit.test.ts
@@ -39,6 +39,7 @@ import BlastRouterProxyFactory_v2_1_0_0 from "../../src/assets/strats/v2.1.0-0/B
 import OrbitLogic_v2_1_0_1 from "../../src/assets/strats/v2.1.0-1/OrbitLogic.json";
 // v2.1.0-6
 import BlastSmartKandelSeeder_v2_1_0_6 from "../../src/assets/strats/v2.1.0-6/BlastSmartKandelSeeder.json";
+import SmartKandelSeeder_v2_1_0_6 from "../../src/assets/strats/v2.1.0-6/SmartKandelSeeder.json";
 // v2.1.0-7
 import UniswapV3Manager_Monoswap_v2_1_0_7 from "../../src/assets/strats/v2.1.0-7/UniswapV3Manager-Monoswap.json";
 import UniswapV3RoutingLogic_Monoswap_v2_1_0_7 from "../../src/assets/strats/v2.1.0-7/UniswapV3RoutingLogic-Monoswap.json";
@@ -861,7 +862,10 @@ describe("strats.ts", () => {
           ),
           orbitLogic: undefined,
           zeroLendLogic: undefined,
-          smartKandelSeeder: undefined,
+          smartKandelSeeder: firstVersionDeploymentsToVersionNetworkDeployment(
+            SmartKandelSeeder_v2_1_0_6,
+            "80001",
+          ),
           uniswapV3ManagerMonoswap: undefined,
           uniswapV3ManagerThruster: undefined,
           uniswapV3RoutingLogicMonoswap: undefined,